requirements.txt
cloudinary==1.25.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
whitenoise==5.2.0